feat: :art: Make sure tags are properly posted for devto and linkedin when cross-posting from medium. Format linkedin post to match 1300 character limit. Url from original post brings users to the actual post.
Extract tags from medium article and make sure its formatted properly for devto and linkedin. Truncate the content to have 1200 characters for linkedin, allow 100 characters for title, url and tags. Convert html to formatted content for linkedin with htmlToContentConverter. Remove the query for the RSS URL so the link works correctly.